# A bit of optimization for Richards3D