I started to work on `Richards3D` and I added several scoped timers to the functions called inside `Richards3D`. The current `v3.0` (configured with `-DMATH_OPTIM=true` on my laptop produce this output when the thest **Muntatschini_ref_005** is run

```
+-------------------------------------------------------------------+------------+------------+
| Total CPU time elapsed since start                                |      86.9s |            |
|                                                                   |            |            |
| Section                                               | no. calls |  CPU time  | % of total |
+-------------------------------------------------------+-----------+------------+------------+
| BiCGSTAB_strict_lower_matrix_plus_identity_by_vector  |        28 |      0.25s |       0.3% |
| EnergyBalance                                         |        24 |     17.38s |      20.0% |
| PointEnergyBalance                                    |    143064 |     17.25s |      19.9% |
| Richards3D                                            |        24 |     56.30s |      64.8% |
| SolvePointEnergyBalance                               |    143064 |      8.19s |       9.4% |
| find_dfdH_3D                                          |        28 |      3.48s |       4.0% |
| find_f_3D                                             |        52 |     12.27s |      14.1% |
| find_matrix_K_3D                                      |        24 |     36.78s |      42.3% |
| get_all_input                                         |         1 |      0.36s |       0.4% |
| product_matrix_using_lower_part_by_vector_plus_vector |        52 |      0.07s |      0.00% |
| product_using_only_strict_lower_diagonal_part         |        52 |      0.06s |      0.00% |
| time_loop                                             |         1 |     86.54s |      99.6% |
| tridiag                                               |        56 |      0.07s |      0.00% |
| tridiag2                                              |    476659 |      0.13s |       0.1% |
| water_balance                                         |        24 |     56.34s |      64.8% |
| write_output                                          |        24 |      0.54s |       0.6% |
+-------------------------------------------------------+-----------+------------+------------+

```
As you can see, the two most expensive functions called by `Richards3D` are `find_matrix_K_3D` and `find_f_3D`. The latter is well suited for a parallelization through openMP. After adding the pragma, and configuring with `-DWITH_OMP=true`, the output is the following (look `find_f_3D`)

```
+-------------------------------------------------------------------+------------+------------+
| Total CPU time elapsed since start                                |      78.2s |            |
|                                                                   |            |            |
| Section                                               | no. calls |  CPU time  | % of total |
+-------------------------------------------------------+-----------+------------+------------+
| BiCGSTAB_strict_lower_matrix_plus_identity_by_vector  |        28 |      0.25s |       0.3% |
| EnergyBalance                                         |        24 |     17.24s |      22.0% |
| PointEnergyBalance                                    |    143064 |     17.11s |      21.9% |
| Richards3D                                            |        24 |     47.55s |      60.8% |
| SolvePointEnergyBalance                               |    143064 |      8.19s |      10.5% |
| find_dfdH_3D                                          |        28 |      3.59s |       4.6% |
| find_f_3D                                             |        52 |      3.25s |       4.2% |
| find_matrix_K_3D                                      |        24 |     36.82s |      47.1% |
| get_all_input                                         |         1 |      0.37s |       0.5% |
| product_matrix_using_lower_part_by_vector_plus_vector |        52 |      0.10s |       0.1% |
| product_using_only_strict_lower_diagonal_part         |        52 |      0.10s |       0.1% |
| time_loop                                             |         1 |     77.84s |      99.5% |
| tridiag                                               |        56 |      0.07s |      0.00% |
| tridiag2                                              |    476659 |      0.13s |       0.2% |
| water_balance                                         |        24 |     47.59s |      60.9% |
| write_output                                          |        24 |      0.59s |       0.8% |
+-------------------------------------------------------+-----------+------------+------------+
```

lookin at the functions that are called I realized that many functions that were written in a `.cc` file should be declared `inline` inside an header since they are pretty small. With the inlining we gained few percent.

```
# clang-6
+---------------------------------------------+------------+------------+
| Total CPU time elapsed since start          |      76.2s |            |
|                                             |            |            |
| Section                         | no. calls |  CPU time  | % of total |
+---------------------------------+-----------+------------+------------+
| EnergyBalance                   |        24 |     16.22s |      21.3% |
| PointEnergyBalance              |    143064 |     16.09s |      21.1% |
| Richards3D                      |        24 |     46.69s |      61.3% |
| SolvePointEnergyBalance         |    143064 |      7.19s |       9.4% |
| find_matrix_K_3D                |        24 |     36.18s |      47.5% |
| get_all_input                   |         1 |      0.33s |       0.4% |
| water_balance                   |        24 |     46.75s |      61.3% |
| write_output                    |        24 |      0.50s |       0.7% |
+---------------------------------+-----------+------------+------------+

# gcc-8.2.0
+---------------------------------------------+------------+------------+
| Total CPU time elapsed since start          |      76.7s |            |
|                                             |            |            |
| Section                         | no. calls |  CPU time  | % of total |
+---------------------------------+-----------+------------+------------+
| EnergyBalance                   |        24 |     16.70s |      21.8% |
| PointEnergyBalance              |    143064 |     16.57s |      21.6% |
| Richards3D                      |        24 |     46.80s |      61.0% |
| SolvePointEnergyBalance         |    143064 |      7.80s |      10.2% |
| find_matrix_K_3D                |        24 |     36.24s |      47.3% |
| get_all_input                   |         1 |      0.37s |       0.5% |
| water_balance                   |        24 |     46.84s |      61.1% |
| write_output                    |        24 |      0.58s |       0.8% |
+---------------------------------+-----------+------------+------------+

```