@asartori86 2018-11-24
I tried several strategies to speed up `find_matrix_K_3D` without success. @yakopuz, there is the variable `cnt` which is updated several times during an iteration of the for loop. It is not clear to me if at beginning of each iteration `cnt = C * i`, i.e. `cnt` is incremented by the same amount per iteration. If so, we may parallelize also this function quite straightforwardly