@asartori86 2018-11-26
@yakopuz @ecor @cozzini @ElisaBortoli, I see that the `cnt` variable is basically updated `3 times` per iteration, **but** sometimes just 2. Do you know why? can we predict for which iteration the previous one updated `cnt` just twice instead of three times? 

I am talking about the `find_matrix_K_3D` function