@yakopuz 2018-11-26
> @yakopuz @ecor @cozzini @ElisaBortoli, I see that the `cnt` variable is basically updated `3 times` per iteration, **but** sometimes just 2. Do you know why? can we predict for which iteration the previous one updated `cnt` just twice instead of three times?

@asartori86  This is apart that I know little ... I see only:

 /** cnt is the counter of Li Lp Lx (lower diagonal without diagonal) */

I see that for different situations (vertical flow, surface flow, and so on) there are loops in count.
@ecor do you understand better how find_matrix_K_3D works?