@yakopuz 2018-11-27
Alberto,

The physics should work in this way:
The model solves before the energy balance, then the water balance.

However, there is some coupling among the two. When you solve the energy balance you calculate the evapotranspiration on the basis of the available water, so you need to subtract this water from the water balance. The functions that are doing this are update_soil_land and update_soil_channel in energy.balance.cc. If the mass budget calculation is not satisfied, this could trigger some recalculation of energy balance and water balance.

I do not know the exact details of the implementation, however I would start to look on time_loop  in geotop.cc. I would look how time stepping is implemented.
From a quick look, I understand that, after some initializations, at line 311 you run first the energy balance, then if this has success en == 0 you run the water balance, any of them do not have success line 329 if (en != 0 || wt != 0), then you try to calculate them again with a shorter time steps. Notice that you can have different error values for en and wt, depending on what happens.

To better understand the code I would recommend (at least it would be very helpful also for me to better assist you and Elisa):
-       A graphical plot of all the functions dependencies (I did something like this with doxygen many years ago)
-       Following in a debug mode the code time stepping for a simple test. I recommend the test https://github.com/geotopmodel/geotop/tree/v3.0/tests/3D/small_example-channel  for its simplicity and completeness.

Giacomo

From: Alberto Sartori <notifications@github.com>
Sent: Tuesday, 27 November, 2018 09:18
To: geotopmodel/geotop <geotop@noreply.github.com>
Cc: Bertoldi Giacomo <giacomo.bertoldi@eurac.edu>; Mention <mention@noreply.github.com>
Subject: Re: [geotopmodel/geotop] A bit of optimization for Richards3D (#96)


thanks @yakopuz<https://github.com/yakopuz>, and do you know if EnergyBalance and WaterBalance could be overlapped? or they modify the same variables and they have to run sequentially?
—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://github.com/geotopmodel/geotop/pull/96>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AP8hXFHWw50dDicTPKelQcEl_7Lo-7vXks5uzPVCgaJpZM4YxlMD>.[https://github.com/notifications/beacon/AP8hXA9lP_4Xr0Is-7QEpgpf1aOCv66iks5uzPVCgaJpZM4YxlMD.gif]