@ecor 2018-11-27
Hi @yakopuz <https://github.com/yakopuz> @ecor <https://github.com/ecor>
@cozzini <https://github.com/cozzini> @ElisaBortoli
<https://github.com/ElisaBortoli>, @asartori86
<https://github.com/asartori86> .
 find_K_3D_matrix looks to calculate a term, the Lx pointer (the 4th
pointer) ,  which is function of hydraulic conductivity and is part of the
matrix that is solved in the core ODE system within the Newton-like
iterator.

Each term of this quantity Lx is refererred to each cell of the Richards3D
integration domain, cnt is a counter .

In fact, it is called in case of update=TRUE.  So I can see from the code.

cheers
Emanuele





Il giorno lun 26 nov 2018 alle ore 18:07 Giacomo Bertoldi <
notifications@github.com> ha scritto:

> @yakopuz <https://github.com/yakopuz> @ecor <https://github.com/ecor>
> @cozzini <https://github.com/cozzini> @ElisaBortoli
> <https://github.com/ElisaBortoli>, I see that the cnt variable is
> basically updated 3 times per iteration, *but* sometimes just 2. Do you
> know why? can we predict for which iteration the previous one updated cnt
> just twice instead of three times?
>
> @asartori86 <https://github.com/asartori86> This is apart that I know
> little ... I see only:
>
> /** cnt is the counter of Li Lp Lx (lower diagonal without diagonal) */
>
> I see that for different situations (vertical flow, surface flow, and so
> on) there are loops in count.
> @ecor <https://github.com/ecor> do you understand better how
> find_matrix_K_3D works?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/geotopmodel/geotop/pull/96#issuecomment-441718275>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEQ3gSY3Eg8Yk1-cmbPXeWuNMmKQkqxIks5uzB_rgaJpZM4YxlMD>
> .
>


-- 
Emanuele Cordano, PhD
Environmental Engineer / Ingegnere per l' Ambiente e il territorio nr.
3587 (Albo A - Provincia di Trento)
cell: +39 3282818564
email: emanuele.cordano@gmail.com,emanuele.cordano@rendena100.eu,
emanuele.cordano@eurac.edu
PEC: emanuele.cordano@ingpec.eu
URL: www.rendena100.eu
LinkedIn: https://www.linkedin.com/in/emanuele-cordano-31995333
GitHub: https://github.com/ecor